# ranking_service/internal/server/routes.go
package server

import (
	"net/http"
	"ranking_service/internal/ranking"
)

func (s *Server) RegisterRoutes() http.Handler {
	mux := http.NewServeMux()

	// Route để kết nối WebSocket
	mux.HandleFunc("/api/videos/ranking/global", ranking.FetchRanking)
	mux.HandleFunc("/api/videos/ranking/user", ranking.FetchUserRanking)

	// Wrap the mux with CORS middleware
	return s.corsMiddleware(mux)
}

func (s *Server) corsMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		// Set CORS headers
		w.Header().Set("Access-Control-Allow-Origin", "*") // Replace "*" with specific origins if needed
		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS, PATCH")
		w.Header().Set("Access-Control-Allow-Headers", "Accept, Authorization, Content-Type, X-CSRF-Token")
		w.Header().Set("Access-Control-Allow-Credentials", "false") // Set to "true" if credentials are required

		// Handle preflight OPTIONS requests
		if r.Method == http.MethodOptions {
			w.WriteHeader(http.StatusNoContent)
			return
		}

		// Proceed with the next handler
		next.ServeHTTP(w, r)
	})
}